ParticleFile: Move ParticleFile into a separate file
This requires a minor touch-up to the write function to
break a dependency chain.
