+++ parcels/__init__.py
@@ -4,3 +4,4 @@ from parcels.field import *  # NOQA
 from parcels.kernel import *  # NOQA
 import parcels.rng as random  # NOQA
 from parcels.kernels.advection import *  # noqa
+from parcels.particlefile import *  # noqa
